Fix profile data JSON duplicate download issue
The settings section supports primary and secondary actions. And the action passed in is set to the item(bottom bar) and to the button's onClick action. This causes the callback to be fired twice when clicked on the button area. Since there's no real use case to support two separate actions at the same time, the action was removed from the button components and was just added to the bottom bar.

Resolves wso2/identity-apps#98