feat(auth-portal): Reusable countdown, modal, and timeout components
Contains 3 files which are timeout.jsp, countdown.jsp and modal.jsp

countdown.jsp is a pure script file that has a class named
Countdown for implementing robust timeout and countdown
use cases.

timeout.jsp is a component that uses a countdown to redirect
the user back to sign-in page when it hits a timeout.

modal.jsp is a component that can be reused in other components to show a tiny
modal. Dev can pass a dynamic title, a description, cancel button
action, and a primary button action to the component via parameters or
by using the ModalRef class.