refactor(modules): Add form imperative event handlers
With this change the Form component uses a forward reference to
expose submit handler to the consumer parent component.

Since, we have the submit button externally we can use the method
`triggerSubmit` to synthetically dispatch the submit event of the
form element. The consumer have to keep a ref of the form to invoke
this handle via `ref.current`.

Also, this change was introduced because Wizard implementation only
facilates the onSubmit() for the form and hence triggering only
the last step. Causing react-final-form to not work as expected.