feat(modules): Implement strategy based file picker component
Contains a strategy based file picker component that can upload
files and paste file contents. This is similar to the upload
certificate component.

However, this implementation is more dynamic and it uses the
strategy pattern to support more than one file type. It
supports serialization, and validation as for the basic
functionality.