Fix validation issues on SAML and OIDC forms.
Both forms can handle the validations synchronously
accordingly. Now it preserves the state while navigating
back and fourth inside the wizard component.

A known issue is that, when you enter a invalid value to
a field and navigate back to a previous page and return
back, it does not focus the error field. The next button
is disabled anyways but the field should be focused to
indicate that its a erroneous value. Currently, the only
way to get the message is by either editing or focusing the
field manually.