fix(console): impl of new resource name validate function
With this new validation function we strictly check the allowed characters in
IdPs and SPs resource names. For example, the function allows inputs like
below: -

   `{ SAMPLE_NAME, sample-name, Sample-Name, Sample_Name, sample_name, My App, My.App }`

but it will reject names like: -

   `{ 1Sample-Name, _Sample, -MySample_Name-1 }`

In laymen terms, it only allows names which only contains alphabetic characters
either simple case or in caps including numbers and special characters `dash` and
`dot` and `underscore`. Users aren't allowed to start resource names with numbers
or special characters in any way.

In this commit: -

- We have updated the validation function of `resourceName`
  and replaced it with the new one.