feat(console): Revamp of IdP Certificates upload and JWKS input form (#2376)
* fix(i18n): change IdP cert modal title

* fix(react-components): hint text line alignment

Now the default hint is an inline-flex element.
This change is applied directly within the
React component itself.

* fix(console): update general details form

* chore(console): remove obsolete code block

* feat(console): pass down saml prop

* feat(console): implement new cert component

* fix(i18n): update cert title in auth providers modal

* chore(console): correct misleading doc comment.

* fix(console): description update of E-IdPs 5761

* fix(console): toggle default cert config mode

Now based on certificate configuration mode; JWKS or Certificates the component
now select that option and toggle it automatically.

if (not(jwks)) then
   show(certificates)
elseif (not(certificates)) then
   show(jwks)
end

* fix(react-components): impl switch disable the feature

Now you can disable a switch individually and give a common
disabled message via component props.

* fix(console): feature to disable JWKS switch on wizard and form

* fix(react-components): resolve initial toggle state in `Switcher`

Removed the default selected value. It was a confusing value
to work with. Now it simply expects `selectedValue` optionally
if you want to override the currently selected value.

* fix(console): enable JWKS for OIDC

* review: remove inline styles

* review: remove `marking` comments in between

* review: move constants to the top

* review: change licence copyright year