Fix validation issues of IdPs and wizard UX improvements (#2397)
* fix(i18n): change SSOUrl field labels

Changed the SSOUrl label in both wizard and
authenticator form component.

* fix(console): switch the order of fields

Now required but already filled fields are moved
to the bottom of the wizard form.

* fix(console): re-order the help-panel descriptions

* fix(console): impl of new resource name validate function

With this new validation function we strictly check the allowed characters in
IdPs and SPs resource names. For example, the function allows inputs like
below: -

   `{ SAMPLE_NAME, sample-name, Sample-Name, Sample_Name, sample_name, My App, My.App }`

but it will reject names like: -

   `{ 1Sample-Name, _Sample, -MySample_Name-1 }`

In laymen terms, it only allows names which only contains alphabetic characters
either simple case or in caps including numbers and special characters `dash` and
`dot` and `underscore`. Users aren't allowed to start resource names with numbers
or special characters in any way.

In this commit: -

- We have updated the validation function of `resourceName`
  and replaced it with the new one.

* fix(console): update validation of IdP names

Now for the IdP name validation we use the new
function `isValidResourceName` instead of `resourceName`
function.

* fix(console): update attribute name validation

* review: refer error message from a constant