fix(applications): re-initialize application templates on refresh (#2719)
* fix(applications): re initialize application templates.

When navigating to an application using the direct url i.e.,
/t/foo/develop/applications/:id#tab=1 this component will be
mounted. But you see this {@link applicationTemplates}?; it is
loaded elsewhere. For some reason, requesting the state from
{@link useSelector} always returns {@code undefined} when
directly navigating or refreshing the page. Therefore; it hangs
without doing anything. It will show a overlay loader but
that's it. Nothing happens afterwards.

So, as a workaround; if for some reason, the {@link useSelector}
return no data, we will manually emit the event
{@link ApplicationActionTypes.SET_APPLICATION_TEMPLATES}. So, doing
that ensures we load the application templates again.

Consider this as a **failsafe workaround**. We shouldn't rely
on this. This may get removed in the future.

* review: remove setTimeout wrapper